app/requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
django-allauth==0.51.0
Django==4.0.4
gunicorn==20.1.0
python-decouple==3.6
sqlparse==0.4.2
psycopg2-binary==2.9.1